DDD-16 Move React files to external folder
