--- src/main/webapp/src/App.js
@@ -1,5 +1,0 @@
-function App() {
-  return <div></div>;
-}
-
-export default App;
